Use uint32_t instead of uint16_t for the size.  From a practical standpoint, no one's probably going to legitimately use the full extent of either the short or the long for a point format size though

+++ include/liblas/lasformat.hpp
@@ -90,7 +90,7 @@ public:
   
 protected:
     
-    liblas::uint16_t m_size;
+    liblas::uint32_t m_size;
     liblas::uint8_t m_versionminor;
     liblas::uint8_t m_versionmajor;
